+++ src/components/home/walking/EmptyWalkingMoneyRoad.tsx
