+++ src/components/home/walking/ProceedingStemp.stories.tsx
