+++ src/components/home/walking/ProceedingStemp.tsx
