+++ src/components/home/walking/WalkingMoneyRoadItem.stories.tsx
