+++ src/pages/Home/KidCreate.tsx
@@ -1,14 +1,14 @@
 import { useParams, Navigate } from 'react-router-dom';
 import styled from 'styled-components';
-import Step1 from '@components/home/kid/create/steps/Step1';
-import Step2 from '@components/home/kid/create/steps/Step2';
-import Progress from '@components/home/kid/create/Progress';
+import Step1 from '@components/home/create/steps/Step1';
+import Step2 from '@components/home/create/steps/Step2';
+import Progress from '@components/home/create/Progress';
 import MarginTemplate from '@components/layout/MarginTemplate';
-import Step3 from '@components/home/kid/create/steps/Step3';
-import Step4 from '@components/home/kid/create/steps/Step4';
+import Step3 from '@components/home/create/steps/Step3';
+import Step4 from '@components/home/create/steps/Step4';
 import { useEffect, useState } from 'react';
 import useAxiosPrivate from '@hooks/auth/useAxiosPrivate';
-import Step5 from '@components/home/kid/create/steps/Step5';
+import Step5 from '@components/home/create/steps/Step5';
 import { useAppDispatch } from '@store/app/hooks';
 import { dispatchParent } from '@store/slices/challengePayloadSlice';
 import { kidMock } from '@lib/mocks/kid';
